improved-diffusion/requirements.txt
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
blis==1.2.0
blobfile==3.0.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
comm==0.2.2
confection==0.1.5
cymem==2.0.11
debugpy==1.8.13
decorator==5.2.1
docker-pycreds==0.4.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.17.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
-e git+https://github.com/chuiyang/tgm-dlm.git@fdbda6818f1e7aa3d0dc73b36e3d5baca4fae1b1#egg=improved_diffusion&subdirectory=improved-diffusion
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpi4py==4.0.3
mpmath==1.3.0
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
rdkit==2024.9.5
regex==2024.11.6
requests==2.32.3
rich==13.9.4
sacremoses==0.1.1
sentry-sdk==2.22.0
setproctitle==1.3.5
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
thinc==8.3.4
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
-e git+https://github.com/chuiyang/tgm-dlm.git@fdbda6818f1e7aa3d0dc73b36e3d5baca4fae1b1#egg=transformers&subdirectory=transformers
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.8
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
zipp==3.21.0
